Resolve S3 credentials via the full AWS provider chain
