+++ apiserver/src/main/resources/application.properties
@@ -1138,10 +1138,15 @@ dt.file-storage.local.directory=${dt.data-directory}/storage
 # `dt.file-storage.s3.secret-key` are used. When both are missing, requests are performed
 # anonymously. When only one of the two is set, startup fails.
 # <br/><br/>
-# When set to `aws`, credentials are resolved from the AWS credential chain. This supports
-# ECS task roles, IRSA / web identity tokens, and the EC2 instance metadata service, and
-# does not require static keys to be configured. `dt.file-storage.s3.access-key` and
-# `dt.file-storage.s3.secret-key` are ignored in this mode.
+# When set to `aws`, credentials are resolved from the environment, mirroring the credential
+# provider chain of the AWS SDK: `AWS_*` environment variables, the shared AWS config file
+# (~/.aws/credentials), ECS task roles, EKS IRSA / web identity tokens, and the EC2 instance
+# metadata service, in that order. Static keys are not required and are ignored in this mode.
+# Startup fails if no credentials can be resolved.
+# <br/><br/>
+# EKS Pod Identity is not supported: the S3 client does not read
+# AWS_CONTAINER_AUTHORIZATION_TOKEN_FILE and rejects the non-loopback Pod Identity Agent
+# endpoint. Use IRSA instead.
 #
 # @category:     Storage
 # @default:      static
