+++ file-storage/provider-s3/src/test/java/org/dependencytrack/filestorage/s3/S3FileStorageTest.java
@@ -19,10 +19,7 @@
 package org.dependencytrack.filestorage.s3;
 
 import com.adobe.testing.s3mock.testcontainers.S3MockContainer;
-import io.minio.credentials.IamAwsProvider;
-import io.minio.credentials.StaticProvider;
 import io.smallrye.config.SmallRyeConfigBuilder;
-import okhttp3.OkHttpClient;
 import org.dependencytrack.filestorage.api.FileStorage;
 import org.dependencytrack.filestorage.proto.v1.FileMetadata;
 import org.eclipse.microprofile.config.Config;
@@ -53,12 +50,6 @@ class S3FileStorageTest {
                     .withInitialBuckets("test");
 
     @Test
-    void shouldHaveNameS3() {
-        final var provider = new S3FileStorageProvider();
-        assertThat(provider.name()).isEqualTo("s3");
-    }
-
-    @Test
     void shouldThrowWhenBucketDoesNotExist() {
         assertThatExceptionOfType(IllegalStateException.class)
                 .isThrownBy(() -> {
@@ -83,77 +74,6 @@ class S3FileStorageTest {
                             Map.entry("dt.file-storage.s3.connect-timeout-ms", "500")))) {}
                 })
                 .withMessage("Failed to determine if bucket does-not-exist exists");
-    }
-
-    @Test
-    void resolveCredentialsProviderShouldReturnStaticProviderByDefault() {
-        final Config config = configOf(Map.ofEntries(
-                Map.entry("dt.file-storage.s3.access-key", "foo"),
-                Map.entry("dt.file-storage.s3.secret-key", "bar")));
-
-        assertThat(S3FileStorageProvider.resolveCredentialsProvider(config, new OkHttpClient()))
-                .hasValueSatisfying(provider -> assertThat(provider).isInstanceOf(StaticProvider.class));
-    }
-
-    @Test
-    void resolveCredentialsProviderShouldReturnStaticProviderWhenSourceIsStatic() {
-        final Config config = configOf(Map.ofEntries(
-                Map.entry("dt.file-storage.s3.credentials-source", "static"),
-                Map.entry("dt.file-storage.s3.access-key", "foo"),
-                Map.entry("dt.file-storage.s3.secret-key", "bar")));
-
-        assertThat(S3FileStorageProvider.resolveCredentialsProvider(config, new OkHttpClient()))
-                .hasValueSatisfying(provider -> assertThat(provider).isInstanceOf(StaticProvider.class));
-    }
-
-    @Test
-    void resolveCredentialsProviderShouldReturnEmptyWhenStaticKeysAreMissing() {
-        final Config config = configOf(Map.ofEntries(
-                Map.entry("dt.file-storage.s3.credentials-source", "static")));
-
-        assertThat(S3FileStorageProvider.resolveCredentialsProvider(config, new OkHttpClient())).isEmpty();
-    }
-
-    @Test
-    void resolveCredentialsProviderShouldThrowWhenOnlyAccessKeyIsConfigured() {
-        final Config config = configOf(Map.ofEntries(
-                Map.entry("dt.file-storage.s3.credentials-source", "static"),
-                Map.entry("dt.file-storage.s3.access-key", "foo")));
-
-        assertThatExceptionOfType(IllegalStateException.class)
-                .isThrownBy(() -> S3FileStorageProvider.resolveCredentialsProvider(config, new OkHttpClient()))
-                .withMessage("Both dt.file-storage.s3.access-key and dt.file-storage.s3.secret-key "
-                        + "must be set when using static credentials");
-    }
-
-    @Test
-    void resolveCredentialsProviderShouldThrowWhenOnlySecretKeyIsConfigured() {
-        final Config config = configOf(Map.ofEntries(
-                Map.entry("dt.file-storage.s3.credentials-source", "static"),
-                Map.entry("dt.file-storage.s3.secret-key", "bar")));
-
-        assertThatExceptionOfType(IllegalStateException.class)
-                .isThrownBy(() -> S3FileStorageProvider.resolveCredentialsProvider(config, new OkHttpClient()))
-                .withMessage("Both dt.file-storage.s3.access-key and dt.file-storage.s3.secret-key "
-                        + "must be set when using static credentials");
-    }
-
-    @Test
-    void resolveCredentialsProviderShouldReturnIamAwsProviderWhenSourceIsAws() {
-        final Config config = configOf(Map.ofEntries(
-                Map.entry("dt.file-storage.s3.credentials-source", "aws")));
-
-        assertThat(S3FileStorageProvider.resolveCredentialsProvider(config, new OkHttpClient()))
-                .hasValueSatisfying(provider -> assertThat(provider).isInstanceOf(IamAwsProvider.class));
-    }
-
-    @Test
-    void resolveCredentialsProviderShouldThrowWhenSourceIsInvalid() {
-        final Config config = configOf(Map.ofEntries(
-                Map.entry("dt.file-storage.s3.credentials-source", "bogus")));
-
-        assertThatExceptionOfType(IllegalArgumentException.class)
-                .isThrownBy(() -> S3FileStorageProvider.resolveCredentialsProvider(config, new OkHttpClient()));
     }
 
     @Test
@@ -284,12 +204,6 @@ class S3FileStorageTest {
                 .withDefaultValues(configValues)
                 .build();
         return new S3FileStorageProvider().create(config, ProxySelector.getDefault());
-    }
-
-    private static Config configOf(Map<String, String> configValues) {
-        return new SmallRyeConfigBuilder()
-                .withDefaultValues(configValues)
-                .build();
     }
 
 }
